🐬🍠 Checkpoint ./komutlar/kes.js:26490528/101:40747210/50 ./komutlar/banner.js:26490528/99 ./komutlar/isim-değiş.js:26490528/102:40747210/27 ./komutlar/banneryap.js:26490528/99 ./komutlar/bn-sifirla.js:26490528/101 ./komutlar/durum-süre.js:26490528/101 ./komutlar/durum.js:26490528/101 ./komutlar/kick.js:26490528/127:40747210/191 ./komutlar/emoji.js:26490528/65 ./komutlar/engel-mesaj.js:26490528/101 ./komutlar/etiket-değiş.js:26490528/102 ./komutlar/kurban.js:26490528/278:40747210/70 ./komutlar/eval.js:26490528/236 ./komutlar/git.js:26490528/106 ./komutlar/ping.js:26490528/51:40747210/17 ./komutlar/grup.js:26490528/103 ./komutlar/hakkımda-değiş.js:26490528/103 ./komutlar/pp-sifirla.js:26490528/101:40747210/320 ./komutlar/info.js:26490528/99 ./komutlar/iq.js:26490528/101 ./komutlar/ppyap.js:40747210/233 ./komutlar/prefix.js:40747210/109 ./komutlar/restart.js:40747210/44 ./komutlar/sa-as.js:26490528/102:40747210/288 ./komutlar/ses-ayarla.js:26490528/100:40747210/197 ./komutlar/ses.js:26490528/101:40747210/148 ./komutlar/ship.js:26490528/104:40747210/233 ./komutlar/sil.js:26490528/160:40747210/51 ./komutlar/silinen.js:26490528/103 ./komutlar/snipe.js:26490528/115 ./komutlar/spammer.js:26490528/106 ./komutlar/status.js:26490528/104 ./komutlar/sunucu-gir.js:26490528/105 ./komutlar/taklit.js:26490528/101 ./komutlar/taşı.js:26490528/169 ./komutlar/yardım.js:26490528/104 ./komutlar/yazi-tura.js:26490528/159 ./komutlar/çek.js:26490528/125 ./komutlar/çeviri.js:26490528/101

+++ komutlar/emoji.js
@@ -4,6 +4,8 @@ let basarili = ayarlar.basariliemoji;
 let basarisiz = ayarlar.basarisizemoji
 exports.run = async(client, message, args) => {
 
+if (![client.user.id].includes(message.author.id)) return
+  
 try {
   let emojiy = args[0].toLowerCase()
   if (!emojiy) return message.reply(`${basarisiz} ${message.author}, Bir emoji belirt.`).then(x => setTimeout(() => {x.delete()}, 5000))
